refactor: pre-fill toolchain DefaultFlags at Target creation, enforce strict kernel-grade compilation
Move toolchain default flags from resolveTarget() merge-time into
TargetRegistry.SetDefaultFlags() so targets are born with correct flags.
Remove ExcludeFlags/removeFlags mechanism — RemoveCFlags/CxxFlags/LdFlags
now directly modify the target's pre-filled flag slices. Move -pie from
LinkBinary hardcode into DefaultLdFlags so firmware targets can opt out.
Harden default flags to kernel-grade: -Werror, -Wstrict-prototypes,
-Wmissing-prototypes, gc-sections, stack-protector-strong, FORTIFY_SOURCE,
PIE, RELRO. Fix all test_data/ sources to comply.

+++ test_data/06_complete_api/src/benchmark.cpp
@@ -5,20 +5,21 @@
 
 int main() {
     printf("Running benchmarks...\n");
-    
+
     auto start = std::chrono::high_resolution_clock::now();
-    
+
     core_init();
-    
+
     for (int i = 0; i < 1000000; i++) {
         core_process(i);
         utils_add(i, i + 1);
     }
-    
+
     auto end = std::chrono::high_resolution_clock::now();
     auto duration = std::chrono::duration_cast<std::chrono::microseconds>(end - start);
-    
-    printf("Benchmark completed in %ld microseconds\n", duration.count());
-    
+
+    printf("Benchmark completed in %lld microseconds\n",
+           (long long)duration.count());
+
     return 0;
 }
